State mgt file organized, react native updated
