+++ app/_layout.tsx
@@ -4,6 +4,8 @@ import { useFonts } from 'expo-font';
 import { SplashScreen, Stack } from 'expo-router';
 import { useEffect } from 'react';
 import { useColorScheme } from 'react-native';
+import { Provider} from 'react-redux';
+import store from '../state mgt/store/store';
 
 export {
   // Catch any errors thrown by the Layout component.
@@ -46,11 +48,14 @@ function RootLayoutNav() {
   const colorScheme = useColorScheme();
 
   return (
+    <Provider store={store}>
     <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
       <Stack>
         <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
         <Stack.Screen name="modal" options={{ presentation: 'modal' }} />
       </Stack>
     </ThemeProvider>
+
+    </Provider>
   );
 }
