+++ state mgt/store/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import userReducer from './userSlice';
+import userReducer from '../slices/userSlice';
 
 const store = configureStore({
   reducer: {
